TCU: removed can_pf field from XlateCoreReq.
This field was used to notify PEMux whether a PF is allowed for the
translation or not and was required when READ and WRITE commands could
specify whether PFs should abort the transfer or not. However, now that
we can only access remote physical memory instead of virtual memory,
this option does no longer exist making also the can_pf field obsolete.