dev: kecacc: Add simple backend for Keccak/SHA-3 accelerator using XKCP
The backend can be used outside of gem5 as well since it does not make
use of any gem5-specific APIs. It is mainly a simple wrapper around the
XKCP (eXtended Keccak Code Package) with handling of the core accelerator
functionality (e.g. absorbing incomplete blocks, initializing based on
the hash type etc).

Change-Id: I0d97f92cfd66008141044bf1b82d04ec344cd28e