Replaced kecacc-xkcp with symlinks to the M³ repo.
This is suboptimal, because it means that gem5 cannot be build without
being embedded within the M³. However, it is better than the other way
around, because we can at least build M³ for host without loading the
gem5 submodule. After all, this repo is only useful within M³ anyway.