dev: fixed writeback bug in i8254xGBe model.
Since the descriptor cache might have changed in the meantime, we always
want to revisit the check to writeback something. However, there can
also be multiple writeback attempts while a writeback is running, so
that we should remember the least restrictive writeback for the retry.
I am not sure if that's 100% correct, though.