TCU: extended PRINT register to write out code coverage info.
Apparently, functional memory accesses are buggy and do, for example,
yield wrong data if a cache eviction is currently in progress. Since
there is no easy way to work around that issue, I decided to add
specific functionality into the TCU to write-out the code coverage
information instead of using gem5_writefile, which relies on functional
memory accesses.

The basic idea is that the software specifies the virtual address, size,
and activity id of the code coverage information to output and the TCU
will then translate the virtual addresses, load the data from memory
cacheline by cacheline, and write it into
<out>/coverage-<tile>-<activity>.profraw.