TCU: removed last fractions for atomic mode.
After thinking about that for quite some time and even trying to add
some support for atomic mode, I've decided to remove it. In more detail:

1. Implementing only the atomic protocol in the TCU is probably a bad
idea, because we would drop all simulation of contention and thereby
lose quite some accuracy.

2. Using the atomic protocol only in parts of the TCU is also worth
thinking about, because the atomic protocol is much simpler than the
timing protocol. However, almost all of the complexity comes from
memory requests using the timing protocol and thereby becoming
asynchronous. But switching to atomic there would basically make the
entire TCU atomic, because memory requests are sent over the NoC in
case of cache misses and thereby it covers pretty much everything of
the TCU.

3. Supporting both the atomic and timing protocol would increase the
implementation complexity even further due to their vast differences.
There does not seem to be a way to abstract their differences away,
because atomic mode is synchronous, whereas timing is asynchronous.
Therefore, supporting both probably means to have a "do-it" function
and two functions that call the "do-it" function for every little step
in every feature we have (many dozens).

The only reasonable way to move forward seems to be to keep using the
timing protocol exclusively, which we need to support the O3 CPU.
Having support for atomic would be nice for some simulation speedup and
potentially also KVM support though. However, I don't see a reasonable
way to get there. Therefore, I've now removed the last fractions of
code that pretended the TCU supports atomic mode.