TCU: added support for multiple chips.
The hardware now uses a 14-bit id, consisting of a 6-bit chip id and a
8-bit tile id. In gem5 (both Python and C++), we represent that with
the TileId struct, containing these two fields. This keeps the exact
encoding at a single place and prevents mistakes.

We also use new names for tiles in the form of C<cid>T<tid> to keep the
ids human readable.