TCU: drain the CPU before stopping it during reset.
In some cases, the kernel is faster to reset the core before the core
can execute the wait-for-interrupt instruction. Since the core, for
example, be stuck in a memory access, we cannot just suspend the core
and resume later, because it will then still be stuck in that memory
access. Instead, we need to properly drain it first to ensure that we
can use it again in a fresh state later.