mem: scratchpad: Drop TODO
in_addr_map controls if the memory address range is considered to be
part of the System physical memory map (here: per tile). It's always set
to true already in tcu_fs.py so we can just change the default.

conf_table_reported controls if the the memory is "reported" to the
kernel e.g. via ACPI. This isn't relevant for M3 but for consistency
with the DRAM setup it can just stay enabled.

Change-Id: Ieae3338b4b1808b1289c0cfb1ce7aea7f7dbea18