mem: scratchpad: Make handling invalid accesses configurable
Unconditionally handling invalid accesses for the entire address space
makes it impossible to configure two or more separate scratchpads/SRAMs
for a tile because they have conflicting address ranges.

It seems like this behavior was introduced in commit 5f71eafefd12
("DTU: connect the cache to a memory-PE.") to handle invalid speculative
memory accesses from the CPU. It does not feel entirely right to make
the Scratchpad handle that - the CPU should not do invalid speculative
accesses in the first place if configured correctly, or they should be
ignored somewhere else in the memory hierarchy. It seems like other
examples in gem5 use a "BadAddr responder" for this (defined in
src/dev/Device.py).

Nevertheless, keep the current behavior as an option for now and use it
by default for tiles with internal scratchpad to avoid changing the
behavior.

Change-Id: I5511f4b06d5eca0e83029ae4d3cd8389a239d7bd