mem: scratchpad: Handle functional accesses correctly
When using Scratchpad as Boot ROM using writeable=False, all write
accesses will be ignored. However, during simulation startup we still
want to load the binary into the Boot ROM. Normally, the writeable check
is bypassed for the functional accesses during simulation startup, but
the Scratchpad does not implement those correctly at the moment.

Fix this by overriding the recvFunctional() implementation of the
SimpleTimingPort and pass those requests directly to functionalAccess()
of AbstractMemory. It's a bit of a hack since SimpleTimingPort is
defined as "a queued port that implements recvFunctional and
recvTimingReq through recvAtomic". The alternative would be to skip
using SimpleTimingPort completely and rely directly on the
QueuedResponsePort (duplicating the parts of SimpleTimingPort).

Change-Id: I56300023afd06dbf8bdc2194c47d6f7083a95119