TCU: Use IOXBar for tiles without caches.
L2XBar claims that its latencies are covered by the caches, which won't
be the case for tiles that do not have any caches. Also, it does not
really make sense to have a CoherentXBar inside a tile without caches.

IOXBar fits a bit better, since it's a NoncoherentXBar and it properly
models some latencies.

IOXBar has a different default width compared to L2XBar (128-bit vs
256-bit). Actually none of the two make sense by default since the CPU
makes accesses in terms of the "cache line size" (64 bytes by default).
The Scratchpad also has a throughput of 512-bit/cycle by default.

While it would be quite expensive to have a 512-bit XBar in practice,
make that the default for consistency so the CPU can actually make its
accesses properly. More specialized tiles (like with the hash
accelerator/KecAcc) should customize this with a more realistic setup.

Change-Id: Iba050fc321cd7f10bed6280aefdf2d2ae3db0f4a