TCU: fixed bug in abort handling.
We cannot abort messages that have already been received at the other
side. So far that was possible if the abort was triggered after we
received the NoC response from the other side, but before the command
was finally finished. In this case, we considered the command to be at
our side and thus aborted it with a non-zero abort result, causing M³ to
repeat the command later (which of cause breaks things as messages
typically cannot be sent twice).