TCU: wait for EP writeback before finishing commands.
Previously, if EP writeback took a long time, it could happen that the
command was already finished and thus the core issued a new command.
This used again the same EPCache and therefore potentially scheduled the
event within this EPCache. As the writeback is still running, the event
might have already been scheduled, leading to a crash (and potentially
other weird effects since we are using the same EPCache simultaneously).

This also required to delay the destruction of the XferUnit::Event in
some cases to ensure that it is still alive when the EP writeback is
finished. In such cases, we destruct it manually in onFinished.